pyproject.toml
[project]
name = "flaskproject"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "ariadne>=0.26.2",
    "bcrypt>=5.0.0",
    "flask>=3.1.2",
    "flask-jwt-extended>=4.7.1",
    "neo4j>=6.0.2",
    "python-dotenv>=1.1.1",
]
